backend/requirements.txt
fastapi
uvicorn
python-multipart
transformers
torch
sentence-transformers
faiss-cpu
langchain
langchain-community
pypdf